build: turn on all warnings and fix errs

+++ src/core/HandlerChain.cpp
@@ -13,7 +13,7 @@ namespace sa {
 void HandlerChain::chain() {
   assert(!this->Handlers_.empty() && this->Handlers_[0].get() &&
          "the handlers in Handler Chain cannot be empty");
-  for (int cur = 0; cur + 1 < this->Handlers_.size(); ++cur) {
+  for (size_t cur = 0; cur + 1 < this->Handlers_.size(); ++cur) {
     this->Handlers_[cur]->setNext(this->Handlers_[cur + 1].get());
   }
 }
